Refactor : Controller 메서드 이전, 메서드 이름 수정
