+++ src/main/java/techcourse/myblog/web/MainController.java
@@ -4,10 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import techcourse.myblog.domain.ArticleRepository;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.domain.UserRepository;
 import techcourse.myblog.domain.UserRequestDto;
 
 import javax.servlet.http.HttpSession;
@@ -15,22 +12,20 @@ import javax.servlet.http.HttpSession;
 @Controller
 public class MainController {
     private final ArticleRepository articleRepository;
-    private final UserRepository userRepository;
 
     @Autowired
-    public MainController(ArticleRepository articleRepository, UserRepository userRepository) {
+    public MainController(ArticleRepository articleRepository) {
         this.articleRepository = articleRepository;
-        this.userRepository = userRepository;
     }
 
     @GetMapping("/")
-    public String index(Model model) {
+    public String moveMainPage(Model model) {
         model.addAttribute("articles", articleRepository.findAll());
         return "index";
     }
 
     @GetMapping("/writing")
-    public String moveWritePage() {
+    public String moveArticleWritePage() {
         return "article-edit";
     }
 
@@ -49,14 +44,7 @@ public class MainController {
     }
 
     @GetMapping("/signup")
-    public String index(UserRequestDto userRequestDto) {
+    public String moveSignUpPage(UserRequestDto userRequestDto) {
         return "signup";
-    }
-
-    @GetMapping("/mypage/{id}")
-    public String moveMyPage(@PathVariable long id, Model model) {
-        User pageUser = userRepository.findUserById(id);
-        model.addAttribute("pageUser", pageUser);
-        return "mypage";
     }
 }
